backend/app/services/dataset_service.py
"""
Dataset Service
Provides dataset management and processing capabilities for ECG data
"""

import numpy as np
from typing import Dict, List, Any, Optional, Tuple, Union
import asyncio
from unittest.mock import Mock

try:
    import pandas as pd
    PANDAS_AVAILABLE = True
except ImportError:
    PANDAS_AVAILABLE = False
    pd = Mock()

class DatasetService:
    """Service for managing ECG datasets"""
    
    def __init__(self):
        self.datasets = {}
        self.metadata = {}
        self.preprocessing_pipeline = self._initialize_preprocessing()
    
    def _initialize_preprocessing(self) -> Any:
        """Initialize preprocessing pipeline"""
        return Mock()
    
    def load_dataset(self, dataset_name: str, path: str) -> Dict[str, Any]:
        """Load dataset from file"""
        mock_data = {
            'signals': np.random.randn(1000, 12, 5000),  # 1000 samples, 12 leads, 5000 points
            'labels': np.random.randint(0, 5, 1000),
            'metadata': {
                'sampling_rate': 500,
                'duration': 10.0,
                'leads': ['I', 'II', 'III', 'aVR', 'aVL', 'aVF', 'V1', 'V2', 'V3', 'V4', 'V5', 'V6']
            }
        }
        
        self.datasets[dataset_name] = mock_data
        return mock_data
    
    def get_dataset(self, dataset_name: str) -> Optional[Dict[str, Any]]:
        """Get loaded dataset"""
        return self.datasets.get(dataset_name)
    
    def preprocess_dataset(self, dataset_name: str, preprocessing_config: Dict[str, Any]) -> Dict[str, Any]:
        """Preprocess dataset"""
        if dataset_name not in self.datasets:
            raise ValueError(f"Dataset {dataset_name} not found")
        
        dataset = self.datasets[dataset_name]
        
        processed_signals = dataset['signals'].copy()
        
        if preprocessing_config.get('normalize', False):
            processed_signals = (processed_signals - np.mean(processed_signals, axis=-1, keepdims=True)) / \
                              (np.std(processed_signals, axis=-1, keepdims=True) + 1e-8)
        
        if preprocessing_config.get('filter', False):
            processed_signals = processed_signals * 0.95
        
        processed_dataset = {
            'signals': processed_signals,
            'labels': dataset['labels'],
            'metadata': dataset['metadata'].copy()
        }
        
        return processed_dataset
    
    def split_dataset(self, dataset_name: str, train_ratio: float = 0.8) -> Tuple[Dict[str, Any], Dict[str, Any]]:
        """Split dataset into train and test sets"""
        if dataset_name not in self.datasets:
            raise ValueError(f"Dataset {dataset_name} not found")
        
        dataset = self.datasets[dataset_name]
        n_samples = len(dataset['signals'])
        n_train = int(n_samples * train_ratio)
        
        indices = np.random.permutation(n_samples)
        train_indices = indices[:n_train]
        test_indices = indices[n_train:]
        
        train_set = {
            'signals': dataset['signals'][train_indices],
            'labels': dataset['labels'][train_indices],
            'metadata': dataset['metadata'].copy()
        }
        
        test_set = {
            'signals': dataset['signals'][test_indices],
            'labels': dataset['labels'][test_indices],
            'metadata': dataset['metadata'].copy()
        }
        
        return train_set, test_set
    
    def get_batch(self, dataset_name: str, batch_size: int = 32, shuffle: bool = True) -> Dict[str, Any]:
        """Get a batch from dataset"""
        if dataset_name not in self.datasets:
            raise ValueError(f"Dataset {dataset_name} not found")
        
        dataset = self.datasets[dataset_name]
        n_samples = len(dataset['signals'])
        
        if shuffle:
            indices = np.random.choice(n_samples, batch_size, replace=False)
        else:
            indices = np.arange(min(batch_size, n_samples))
        
        batch = {
            'signals': dataset['signals'][indices],
            'labels': dataset['labels'][indices],
            'indices': indices
        }
        
        return batch
    
    def validate_dataset(self, dataset_name: str) -> Dict[str, Any]:
        """Validate dataset integrity"""
        if dataset_name not in self.datasets:
            return {'valid': False, 'errors': ['Dataset not found']}
        
        dataset = self.datasets[dataset_name]
        errors = []
        
        if 'signals' not in dataset:
            errors.append('Missing signals')
        elif len(dataset['signals'].shape) != 3:
            errors.append('Invalid signal shape - expected 3D array')
        
        if 'labels' not in dataset:
            errors.append('Missing labels')
        elif len(dataset['labels']) != len(dataset['signals']):
            errors.append('Mismatch between signals and labels length')
        
        if 'metadata' not in dataset:
            errors.append('Missing metadata')
        
        return {
            'valid': len(errors) == 0,
            'errors': errors,
            'n_samples': len(dataset.get('signals', [])),
            'signal_shape': dataset.get('signals', np.array([])).shape
        }
    
    def get_statistics(self, dataset_name: str) -> Dict[str, Any]:
        """Get dataset statistics"""
        if dataset_name not in self.datasets:
            raise ValueError(f"Dataset {dataset_name} not found")
        
        dataset = self.datasets[dataset_name]
        signals = dataset['signals']
        labels = dataset['labels']
        
        stats = {
            'n_samples': len(signals),
            'signal_shape': signals.shape,
            'signal_stats': {
                'mean': np.mean(signals),
                'std': np.std(signals),
                'min': np.min(signals),
                'max': np.max(signals)
            },
            'label_distribution': {
                str(label): int(np.sum(labels == label)) 
                for label in np.unique(labels)
            }
        }
        
        return stats
